upload method done, finishing usersettings
