--- apps/app/src/pages/usersetprofile.tsx
@@ -1,3 +1,0 @@
-import UserSetProfile from '@components/UserSetProfile'
-
-export default UserSetProfile
