+++ apps/app/src/pages/usersetting.tsx
@@ -1,0 +1,3 @@
+import UserSetting from '@components/UserSetting'
+
+export default UserSetting
